pyproject.toml
[project]
name = "avatarforge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.4",
    "uvicorn[standard]==0.32.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "sqlalchemy==2.0.36",
    "alembic==1.14.0",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "httpx==0.27.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.12",
    "requests>=2.31.0",
    "pillow>=10.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["avatarforge*"]
exclude = ["backend*", "tests*"]
